Support DB Init
- provistioing default resources
- support convert command
- renamed sync resource object for more generic use

Partial-bug: #1774451

+++ pkg/schema/schema.go
@@ -31,7 +31,7 @@ const (
 
 var sqlTypeMap = map[string]string{
 	ObjectType:  "json",
-	IntegerType: "int",
+	IntegerType: "bigint",
 	ArrayType:   "json",
 	BooleanType: "bool",
 	NumberType:  "float",
